[vfs] Migrate vfs code to utils
This way it can be used in userspace as well. We will eventually update
the initial ramdisk format such that we will not need to search for the
starting program in kernel space. (Ideally, we can just place the user
entry code at the start of the ramdisk and just pass control to that.)
This way, we won't need to parse the vfs in kernel space. This can be
done by a user task.

- Migrate malloc code to libc and support user malloc.
- Abstract out the allocating logic into an Allocator class that only
  receives info about a heap allocation that malloc taps into.
  - Create a bootstrap heap allocation that userboot can use.
- Create a cstdlib.
- Move some math functions out of kernel.h and into MathUtils.h.
- Add a custom argument to UserTask that controls how users can copy
  arguments onto the initial shared user space.
- Update various libraries to use libc malloc instead of kmalloc.